[lattice] fix bug for nw>1, adjust memory approx
The function imtime_bubble_chi0_wk throwed an error for nw>1 at del chi_tr
 because chi_tr was not assigned and should have been chi0_tr.

I tried to give a better estimate for the memory usage because atleast
for nw>1 it was off.
For nw=1 the heavy weights are G(r, \tau) and either G(r, i\omega) or
\chi(\tau) * all active cores.
For nw>1 the heavy weights are G(r, tau) and \chi(r, \tau).
The estimation is now better for nw>1, but still overestimates the total
memory usage.