[rpa] first corrections to matrix rpa formalism
This seems to be the correct implementation of the matrix RPA, which I
only add to be sure that everything for the eliashberg equation is
correct.
So its sole purpose is testing.
It follows https://arxiv.org/abs/cond-mat/0309575 WITH ONE EXCEPTION.
Instead of U * chi as in equation (40) and (41) we
use chi * U.
This is because I think it is the correct way to have the correct
pairing.
From a point of observables this does not change anything.

Doumentation will follow.