[eli] enforce compatible wmeshes in eli product
The function eliashberg_g_delta_g_product will now throw an error if the
wmesh of GF is smaller than the one of delta. This should never
be the case if one uses the solve_eliashberg function which already
enforces this, but this if clause still acts as a safety net for other
usage.

The function eliashberg_product will now throw an error if the wmesh of
Gamma is not atleast twice as big as the one of delta. This ensures that
w-n lies in the wmesh of Gamma and must not be approximated.
In a future refactoring of the implementation with Fourier
transformations this enforcement should be unnescessary.

This changed implementation passes the test from the previous commit.