[eli] improve fft implementation
- The eliashberg_product_fft now only does the mandatory stuff.
Everything else, like determining the constant part of Gamma or Fourier
transforming Gamma is done before. For this helper function were
introduced. The way this is handled could easiliy be changed.

- Using the eliashberg_product_fft with the scipy eigenvalue solver
prints multiple warnings. This is dependent on the number of Matsubara
frequncies / tau points that are given. Up to this point I did not see
any major differences if those warnings occur, but it should be kept in
mind and fixed.

- Add a explicit fft version to solve the eliashberg equation. This
allows for explicitly giving the constant term of Gamma.

- Minor changes in the documentation to be consistent.