[fourier] introduce lattice fourier for general target
Added a template that uses the omp parallelized fourier
transformation for Gfs with a general target in fourier_gf.hpp
This template was then used is gf.cpp and chi_imtime.cpp
where duplicate definitions of a parallelized fourier
transformation existed.

The naming a placement of fourier_gf.hpp, or further
generalizations are up for debate.