[lattice] WIP add memory saving calculation chi0
If one is only interested in a subset of bosonic Matsubara
frequencies for the bare bubble it is overkill to store
chi_tr, which can be too huge to store, and then do the
FFT.
Therefore the FFT over tau is put inside of the loop when
only a given range of frequencies is needed.

PROBLEM: This implementation raises  "WARNING: Direct Fourier
cannot deduce the high-frequency moments of G(tau) due to noise",
which is weird.
The results are identical to the previous implementation which
did not raise this warning.